- Added Message Box when exiting application  - Fixed further issues with the Guitar Pro Parser. Looks better now. Anyhow, big shout out to https://github.com/PhilPotter/gp_parser  - Added Musical Notation view to Drum Tracks  - View can be switched via a button on the according track

+++ Widget_Timeline_Tools.cpp
@@ -2070,7 +2070,7 @@ namespace MIDILightDrawer
 
 			// Create or update single preview bar
 			if (previewBar == nullptr) {
-				previewBar = gcnew BarEvent(snapTick, TickLength, Color::FromArgb(100, ColorStart.R, ColorStart.G, ColorStart.B));
+				previewBar = gcnew BarEvent(snapTick, TickLength, Color::FromArgb(100, (int)ColorStart.R, (int)ColorStart.G, (int)ColorStart.B));
 			}
 			else {
 				previewBar->StartTick	= snapTick;
@@ -2263,7 +2263,7 @@ namespace MIDILightDrawer
 
 			// Create or update single preview bar
 			if (previewBar == nullptr) {
-				previewBar = gcnew BarEvent(snapTick, TickLength, Color::FromArgb(100, StrobeColor.R, StrobeColor.G, StrobeColor.B));
+				previewBar = gcnew BarEvent(snapTick, TickLength, Color::FromArgb(100, (int)StrobeColor.R, (int)StrobeColor.G, (int)StrobeColor.B));
 			}
 			else {
 				previewBar->StartTick = snapTick;
